# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BattOS API
  version: 0.1.0-alpha
  description: |
    Contract for BattOS Mission Control. System and Memory Core operations are
    implemented today. Product, run and dashboard operations define the v0.1
    contract and are implemented incrementally by roadmap phase.
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /cli-tools/{id}/install:
    post:
      tags: [Runtimes]
      operationId: requestCliToolInstall
      x-battos-phase: E2
      summary: Request a governed host install for a registered CLI (pending approval)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "201":
          description: Install request created awaiting approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CliToolInstallEnvelope"
  /cli-tools/{id}/installs:
    get:
      tags: [Runtimes]
      operationId: listCliToolInstalls
      x-battos-phase: E2
      summary: List install requests for a CLI tool, newest first
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Install history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CliToolInstallList"
  /cli-tools/installs/{installId}/approve:
    post:
      tags: [Runtimes]
      operationId: approveCliToolInstall
      x-battos-phase: E2
      summary: Decide an install request; approval executes the command on the host
      parameters:
        - name: installId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CliToolInstallDecisionInput"
      responses:
        "200":
          description: Decision recorded (approved installs run in background)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CliToolInstallEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: BattOS admin token
...
  schemas:
...
    CliToolInstall:
      type: object
      required: [id, cli_tool_id, install_command, status, requested_at]
      properties:
        id: { type: string }
        cli_tool_id: { type: string }
        install_command: { type: string }
        status:
          type: string
          enum: [pending_approval, rejected, running, succeeded, failed]
        reason: { type: string }
        output: { type: string }
        requested_at: { type: string, format: date-time }
        decided_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
    CliToolInstallEnvelope:
      type: object
      required: [install]
      properties:
        install:
          $ref: "#/components/schemas/CliToolInstall"
    CliToolInstallList:
      type: array
      items:
        $ref: "#/components/schemas/CliToolInstall"
    CliToolInstallDecisionInput:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approved, rejected]
        reason: { type: string }
...